pyproject.toml
[project]
name = "kg-gnn-pubmed"
version = "0.1.0"
description = ""
authors = [
    {name = "Morris"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "metapub (>=0.5.12,<0.6.0)",
    "networkx (>=3.4.2,<4.0.0)",
    "numpy (>=1.26,<2.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "sentence-transformers (<=4.1.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "poethepoet",
    "torch-geometric (>=2.6.1,<3.0.0)",
    "datasets (>=3.6.0,<4.0.0)",
    "evaluate (>=0.4.3,<0.5.0)",
    "tensorboard (>=2.19.0,<3.0.0)",
    "accelerate (>=0.25.0,<0.28.0)",
    "tensorboardx (>=2.6.2.2,<3.0.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "transformers (>=4.52.3,<5.0.0)",
    "spacy (>=3.8.7,<4.0.0)",
]

[tool.poe.tasks]
install-ltt = "python3 -m pip install light-the-torch"
run-ltt = "python3 -m light_the_torch install --upgrade torch torchaudio torchvision"
autoinstall-torch-cuda = ["install-ltt", "run-ltt"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
